refactor(chat): extract makeItemDragDrop from inventoryComponent
Phase-1 step 1.2: moves the inventory tree's item link/move drag-and-drop
into inventory-dnd.js as a framework-agnostic make* factory (not a hook —
callable from the current imperative component now, and from
useEffect(() => makeItemDragDrop(...).attachX(...).dispose, deps) during
convert-in-place). It returns three imperative binders:

- attachDragSource(row, absPath) — drag source carrying the absolute path
- attachRowDropTarget(row, { absPath, acceptsDrop }) — hub-row drop target;
  acceptsDrop is a getter because hub status is probed asynchronously
- attachListDropZone(list, rootPrefix) — directory-background drop zone

The pure dropTargetPath/clearAllDropTargets helpers and the link/move
showDropMenu move with it. A faithful behavior-preserving extraction.

Verified: node --check, eslint (0 errors), prettier clean; tsc neutral
(9 pre-existing errors, inventory-dnd.js adds 0). Live-app smoke test
(Vite + headless Chromium against a real daemon) renders 16 inventory rows
with the 5 non-special items draggable and zero dnd/lockdown errors —
confirming all three attach* paths execute during construction.
